Improvements to quadrature.
Since we assume grids are cell-centered, we have interpolate the boundary value. For trapezoid, it uses two points to linearly interpolate the boundary. For Simpsons, it uses three points to fit a polynomial.
Both then use the trapezoid rule to fit account for the cell between the interpolated boundary and nearest cell-center.